misc: move validate_configuration to misc
